refactor: relocate map update service
