+++ services/storyteller/index.ts
@@ -9,7 +9,7 @@ export * from '../dialogueService';
 export * from "./systemPrompt";
 export * from './responseParser';
 export * from './promptBuilder';
-export * from '../mapUpdateService';
+export * from '../cartographer/api';
 export * from '../modelDispatcher';
 export * from '../geminiClient';
 export * from '../apiClient';
